fix v4

+++ src/app/pages/quotes/quotes.page.ts
@@ -1,27 +1,15 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-
-import {
-  IonContent,
-  IonCard,
-  IonCardHeader,
-  IonCardTitle,
-  IonCardSubtitle,
-  IonButton
-} from '@ionic/angular/standalone';
-
+import { IonContent, IonCard, IonCardHeader, IonCardTitle, IonCardSubtitle, IonButton } from '@ionic/angular/standalone';
 import { QuoteDbService } from '../../services/quote-db.service';
 import { QuoteFormComponent } from '../../components/quote-form/quote-form.component';
 
 @Component({
   selector: 'app-quotes',
   templateUrl: './quotes.page.html',
-  styleUrls: ['./quotes.page.scss'],
   standalone: true,
   imports: [
     CommonModule,
-    FormsModule,
     IonContent,
     IonCard,
     IonCardHeader,
@@ -35,14 +23,12 @@ export class QuotesPage {
 
   constructor(public quoteDb: QuoteDbService) {}
 
-  async ionViewWillEnter() {
-    await this.quoteDb.initDB();
+  addQuote(q: any) {
+    this.quoteDb.addQuote(q);
   }
 
-  delete(id: number) {
-    this.quoteDb.deleteQuote(id);
-  }
-  onAddQuote(quote: any) {
-  this.quoteDb.addQuote(quote);
+delete(id?: number) {
+  if (!id) return;
+  this.quoteDb.deleteQuote(id);
 }
 }
